Common code extraction
