+++ pallets/snarcos/src/tests/suite.rs
@@ -1,9 +1,10 @@
 use frame_support::{assert_err, assert_ok, sp_runtime, BoundedVec};
 use frame_system::{pallet_prelude::OriginFor, Config};
+use primitives::snarcos::{ProvingSystem, VerificationKeyIdentifier};
 use sp_runtime::traits::Get;
 
 use super::setup::*;
-use crate::{Error, ProvingSystem, VerificationKeyIdentifier, VerificationKeys};
+use crate::{Error, VerificationKeys};
 
 type Snarcos = crate::Pallet<TestRuntime>;
