+++ primitives/Cargo.toml
@@ -8,6 +8,7 @@ license = "Apache 2.0"
 [dependencies]
 codec = { package = "parity-scale-codec", version = "3.0", default-features = false, features = ["derive"] }
 serde = { version = "1.0", features = ["derive"] }
+strum = { package = "strum_macros", version = "0.24", default-features = false }
 scale-info = { version = "2.0", default-features = false, features = ["derive"] }
 
 sp-api = { default-features = false, git = "https://github.com/Cardinal-Cryptography/substrate.git", branch = "liminal" }
