+++ rust-toolchain
@@ -1,1 +1,1 @@
-nightly-2022-08-12
+nightly-2022-11-16
